Overhaul layout manager (#1)
## Description

This PR is focused on rebuilding the layout manager. Instead of measuring children at every step it now passes the tree two times at most. The first pass calculates whatever it can and propagates calculated sizes upwards from children to parents. The second pass relies on the propagated values to calculate the rest of the children. Algorithm is more thoroughly explained in the code comments.

In addition to the new algorithm it also adds more complex layout calculations for columns, rows and stacks, which allows creating more complex UIs. It also handles RTL layouts, and adds `isRTL` check to `ViewManager`.